Removes the methods range part
This commit removes the part that allows to simply provide a list of
named methods. It is not exact enough as e.g. methods might have the
same name however follow other protocols. It is also a bit redundant to
the information provided with the control factors and these are by far
more suitable to find the right data. This closes #20 for now.