Changes descriptions and element names
This commit changes descriptions removing the fixed reference that the
length has to be given in meters to account for the possibility that
there is now a field that allows to specify a unit the value is given
in.

Further more the commit changes the names of fields which had in their
name a min and max as abbreviation. Now we fully spell this out.