Adds an overview document
The open office document provides an overview of the current vocabulary
and the schema. Further more it provides an insight into how the
vocabulary is interacting with the schema. It also represents the latest
state of the vocabulary and the schema and will help to guide shaping
both components to allow for an efficient user interface.