Moves all details below the aspects to module
The module is now no longer called module but Detail. In that way we
have details that characterize aspects which in turn characterize our
search object. This also support the separation of a core from details
that can come from any ontology.

This closes #93 Move all details to modules
This is related to #90 Idea how the core is separated from the rest