Moves the organizational hierarchy
The organizational hierarchy is a mixed thing and does not fit
well anywhere. While a first idea was to put in on the same level
as the other spheres I think now the best thing is to put it into
the methodological part. It contains anatomical ententies up to
spheres and under the methodological aspect It basically means
what was the focus of the actual study or setup.